+++ modules/tealium_adobevisitor/example/android/app/src/main/kotlin/com/tealium/flutter/modules/tealium_adobevisitor_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.tealium.flutter.modules.tealium_adobevisitor_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}