+++ tealium/android/src/main/kotlin/com/tealium/OptionalModule.kt
@@ -1,0 +1,7 @@
+package com.tealium
+
+import com.tealium.core.TealiumConfig
+
+interface OptionalModule {
+    fun configure(config: TealiumConfig)
+}